refactor by converting monetary system to BigDecimal

+++ src/main/java/com/octopetrus/vendingmachine/coins/CoinType.java
@@ -1,20 +1,22 @@
 package com.octopetrus.vendingmachine.coins;
 
+import java.math.BigDecimal;
+
 public enum CoinType {
 
-//    CENT ("Cent", 0.01, 2.500, 19.05),
-    NICKEL ("Nickel", 0.05, 5.000, 21.21),
-    DIME ("Dime", 0.10, 2.268, 17.91),
-    QUARTER ("Quarter dollar", 0.25, 5.670, 24.26);
-//    HALF ("Half dollar", 0.50, 11.340, 30.61),
-//    DOLLAR ("Dollar", 1.00, 8.1, 26.49);
+//    CENT ("Cent", BigDecimal.valueOf(0.01), 2.500, 19.05),
+    NICKEL ("Nickel", BigDecimal.valueOf(0.05), 5.000, 21.21),
+    DIME ("Dime", BigDecimal.valueOf(0.10), 2.268, 17.91),
+    QUARTER ("Quarter dollar", BigDecimal.valueOf(0.25), 5.670, 24.26);
+//    HALF ("Half dollar", BigDecimal.valueOf(0.50), 11.340, 30.61),
+//    DOLLAR ("Dollar", BigDecimal.valueOf(1.00), 8.1, 26.49);
 
     public final String name;
-    public final double value;
+    public final BigDecimal value;
     public final double weight;
     public final double size;
 
-    CoinType(String name, double value, double weight, double size) {
+    CoinType(String name, BigDecimal value, double weight, double size) {
         this.name = name;
         this.value = value;
         this.weight = weight;
@@ -30,7 +32,7 @@ public enum CoinType {
         throw new IllegalStateException(getIllegalCoinTypeMessage(coin));
     }
 
-    public static double getValue(Coin coin) {
+    public static BigDecimal getValue(Coin coin) {
         for (CoinType coinType : CoinType.values()) {
             if (coinType.weight == coin.weight && coinType.size == coin.size)
                 return coinType.value;
